Issues with depid with year DTG/SWV files
Hey Stacy,

Let me know if I am off here, but from what I can tell, but line 78 is supposed to search for SWV files with the full nc name with year (although in this part of the script it seems like depid variable contains the ncname value). However, this line seemed to search for the truncated name (without year). So I changed it to search for the full name (with year).

Line 72 kept erroring and skipping the rest of the TRY statement due to the TAGON variable not existing, even if this deployment had the full (with year) name to the DTG files. So I added an exist, not sure if this is a good fix but it seems to make the code run alright.

Additionally, a conditional error was added on 79 to trigger the catch statement if no SWV files were found, since d3getcues does not error out if no files are found.

Would be great to get your sense of this problem and if this fix works. Thanks Stacy.